+++ src/main/java/com/example/api_project/pojo/ReturnCheck.java
@@ -63,7 +63,7 @@ public class ReturnCheck implements Serializable {
     /**
      * 入库价格
      */
-    private Long inputPrice;
+    private Float inputPrice;
     /**
      * 实际入库数
      */
@@ -131,7 +131,7 @@ public class ReturnCheck implements Serializable {
     /**
      * 出库价格
      */
-    private Long outputPrice;
+    private Float outputPrice;
     /**
      * 实际出库数
      */
@@ -257,11 +257,11 @@ public class ReturnCheck implements Serializable {
         this.inputPlan = inputPlan;
     }
 
-    public Long getInputPrice() {
+    public Float getInputPrice() {
         return inputPrice;
     }
 
-    public void setInputPrice(Long inputPrice) {
+    public void setInputPrice(Float inputPrice) {
         this.inputPrice = inputPrice;
     }
 
@@ -393,11 +393,11 @@ public class ReturnCheck implements Serializable {
         this.outputPlan = outputPlan;
     }
 
-    public Long getOutputPrice() {
+    public Float getOutputPrice() {
         return outputPrice;
     }
 
-    public void setOutputPrice(Long outputPrice) {
+    public void setOutputPrice(Float outputPrice) {
         this.outputPrice = outputPrice;
     }
