+++ src/main/java/com/example/api_project/pojo/ShopkeeperWarehouse.java
@@ -65,31 +65,31 @@ public class ShopkeeperWarehouse implements Serializable {
      */
     private String description;
     private String modelCode;//型号
-    private Integer priceRetail;//零售价
-    private Integer priceWholesaler;//批发价
-    private Integer priceLatestPurchase;//最近进价
+    private Float priceRetail;//零售价
+    private Float priceWholesaler;//批发价
+    private Float priceLatestPurchase;//最近进价
 
-    public Integer getPriceRetail() {
+    public Float getPriceRetail() {
         return priceRetail;
     }
 
-    public void setPriceRetail(Integer priceRetail) {
+    public void setPriceRetail(Float priceRetail) {
         this.priceRetail = priceRetail;
     }
 
-    public Integer getPriceWholesaler() {
+    public Float getPriceWholesaler() {
         return priceWholesaler;
     }
 
-    public void setPriceWholesaler(Integer priceWholesaler) {
+    public void setPriceWholesaler(Float priceWholesaler) {
         this.priceWholesaler = priceWholesaler;
     }
 
-    public Integer getPriceLatestPurchase() {
+    public Float getPriceLatestPurchase() {
         return priceLatestPurchase;
     }
 
-    public void setPriceLatestPurchase(Integer priceLatestPurchase) {
+    public void setPriceLatestPurchase(Float priceLatestPurchase) {
         this.priceLatestPurchase = priceLatestPurchase;
     }
